% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{choose_axis_type}
\alias{choose_axis_type}
\title{Auto-select the Type of Axis to Draw}
\usage{
choose_axis_type(var)
}
\arguments{
\item{var}{Variable name. Must be a character vector of length 1.}
}
\value{
Character vector of length 1; one of:
\itemize{
\item \code{"yn"} -- if \code{var \%in\% c("fished_yesterday", "weekend")}
\item \code{"day"} -- if \code{var == "day"}
\item \code{"percent"} -- e.g., if \code{var \%in\% c("chinook_comp", "btf_chinook_comp")}
\item \code{"regular"} -- otherwise
}
}
\description{
For use in generalized plotting functions where
the type of variable (and thus axis) needs to change
}
