% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_params}
\alias{count_params}
\title{Count the Number of Parameters in a Model}
\usage{
count_params(fit)
}
\arguments{
\item{fit}{A fitted model object of class \code{\link[stats:lm]{lm}}.}
}
\value{
The number of coefficients in the fitted model plus 1 for the residual standard error.
}
\description{
Count the Number of Parameters in a Model
}
