% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_axis_line}
\alias{draw_axis_line}
\title{Draw an Axis Line with No Ticks or Labels}
\usage{
draw_axis_line(side, col = par("col.axis"))
}
\arguments{
\item{side}{The side of the plot to draw the axis on,
1 = x-axis, 2 = y-axis}

\item{col}{Color, defaults to \code{par("col.axis")}}
}
\description{
Draw an Axis Line with No Ticks or Labels
}
