% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_day_axis}
\alias{draw_day_axis}
\title{Draw an Axis showing Dates}
\usage{
draw_day_axis(fday, lday, by, side = 1, col = par("col.axis"), ...)
}
\arguments{
\item{fday}{The first day (1 = June 1)}

\item{lday}{The last day (1 = June 1)}

\item{by}{The interval to draw axis ticks at}

\item{side}{The side of the plot to draw the axis on,
1 = x-axis, 2 = y-axis}

\item{...}{Optional arguments to be passed to \code{\link[graphics:axis]{graphics::axis()}}}
}
\description{
Draw an Axis showing Dates
}
