% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{draw_yn_axis}
\alias{draw_yn_axis}
\title{Draw an Axis Showing Percentages}
\usage{
draw_yn_axis(side, ...)
}
\arguments{
\item{side}{The side of the plot to draw the axis on,
1 = x-axis, 2 = y-axis}

\item{...}{Optional arguments to be passed to \code{\link[graphics:axis]{graphics::axis()}}}
}
\description{
Draw an Axis Showing Percentages
}
