% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{fit_all_subsets}
\alias{fit_all_subsets}
\title{Fit All Subsets for Multiple Response Variables}
\usage{
fit_all_subsets(global_formulae, fit_data, ...)
}
\arguments{
\item{global_formulae}{A named \code{\link[base:list]{list}} object with elements for \code{effort}, \code{totalcpt}, \code{chinook_comp}, \code{chum_comp}, and \code{sockeye_comp} to be passed to the
\code{formula} argument of \code{\link[=fit_global_model_one]{fit_global_model_one()}} separately.}

\item{fit_data}{A \code{\link[base:data.frame]{data.frame}} object storing the variables for regression fitting.}

\item{...}{Optional arguments passed to \code{\link[=fit_all_subsets_one]{fit_all_subsets_one()}}}
}
\description{
Fit All Subsets for Multiple Response Variables
}
