% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_response}
\alias{get_response}
\title{Obtain the Name of the Response Variable for a Given Model}
\usage{
get_response(fit)
}
\arguments{
\item{fit}{A fitted model object with class \code{\link[stats:lm]{lm}} or \code{"averaging"} (created via \code{\link[MuMIn:model.avg]{MuMIn::model.avg()}}).}
}
\value{
One of:
\itemize{
\item \code{"effort"} (number of drift trips),
\item \code{"total_cpt"} (total salmon catch per trip), or
\item \code{"chinook_comp"} (Chinook salmon proportion composition)
}
}
\description{
Obtain the Name of the Response Variable for a Given Model
}
\note{
Although the models are fitted to response variables
on a transformed scale (\code{"log_effort"}, \code{"log_total_cpt"}, \code{"logit_chinook_comp"}),
the name of the non-transformed variable is returned.
}
