% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_name}
\alias{get_var_name}
\title{Obtain the 'Pretty Name' of a Variable}
\usage{
get_var_name(var)
}
\arguments{
\item{var}{The variable name. Returns \code{NA} if not a valid variable.}
}
\description{
For printing axis labels, etc.
}
