% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_wt}
\alias{get_wt}
\title{Obtain Model Weights given a vector of AIC(c) Values}
\usage{
get_wt(AIC)
}
\arguments{
\item{AIC}{Numeric vector of AIC(c) values from multiple models.}
}
\value{
A numeric vector of model weights.
}
\description{
Obtain Model Weights given a vector of AIC(c) Values
}
