% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverse_transform}
\alias{inverse_transform}
\title{Obtain Function to Return the Inverse of the Scale of the Response Variable}
\usage{
inverse_transform(fit)
}
\arguments{
\item{fit}{A fitted model object with class \code{\link[stats:lm]{lm}}}
}
\value{
A function representing the correct inverse function
to back-transform the response variable. E.g., if the response
is either \code{"effort"} or \verb{"total_cpt}", will return \code{\link[base:Log]{base::exp()}} and
if the response is \code{"chinook_comp"}, will return \code{\link[stats:Logistic]{stats::plogis()}}.
}
\description{
Obtain Function to Return the Inverse of the Scale of the Response Variable
}
