% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{loo_pred}
\alias{loo_pred}
\title{Obtain Leave-one-Out Predictions from a Single Model}
\usage{
loo_pred(fit, lag, fit_data)
}
\arguments{
\item{fit}{A fitted model object of class \code{\link[stats:lm]{lm}}.}

\item{lag}{number of days before the proposed opener the prediction is to be made. See \code{\link[=predict_sonar]{predict_sonar()}}.
Ignored if model does not contain a sonar variable.}

\item{fit_data}{\code{\link[base:data.frame]{data.frame}} storing the variables for regression fitting; generated by \code{\link[KuskoHarvData:prepare_regression_data]{KuskoHarvData::prepare_regression_data()}}}
}
\value{
A \code{\link[base:matrix]{matrix}} object with columns storing values for each record and rows storing the predicted value for
the left out observation and the AICc value (given by \code{\link[MuMIn:AICc]{MuMIn::AICc()}}) of the model fitted after leaving the data point out.
}
\description{
Obtain Leave-one-Out Predictions from a Single Model
}
