% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{loo_pred_model_avg}
\alias{loo_pred_model_avg}
\title{Obtain Model-Averaged Leave-one-Out Predictions}
\usage{
loo_pred_model_avg(fit_list)
}
\arguments{
\item{fit_list}{\code{\link[base:list]{list}} of fitted model objects of class \code{\link[stats:lm]{lm}}, generally the output of \code{\link[=fit_all_subsets_one]{fit_all_subsets_one()}}.}
}
\value{
Numeric vector of model-averaged leave-one-out predictions.
}
\description{
Obtain Model-Averaged Leave-one-Out Predictions
}
\note{
Leave-one-out predictions are made for each model and each record
using \code{\link[=loo_pred]{loo_pred()}} and model weights are recalculated for each left out record
prior to averaging the predictions from each model.
}
