% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{pred_vs_obs}
\alias{pred_vs_obs}
\title{Plot Predicted vs. Observed Points}
\usage{
pred_vs_obs(
  yhat,
  yobs,
  period = NULL,
  xlab = "Observed",
  ylab = "Predicted",
  main = NULL,
  include_summaries = TRUE,
  period_legend = !is.null(period)
)
}
\arguments{
\item{yhat}{Numeric vector of predicted values.}

\item{yobs}{Numeric vector of observed values.}

\item{period}{Optional numeric vector with the period identifier
for each record. Must be one of 1, 2, or 3. If not supplied (default),
all points will be grey. If provided, the points will be color-coded by period.}

\item{xlab}{Character string for x-axis label.}

\item{ylab}{Character string for y-axis label.}

\item{main}{Character string for main plot title.}

\item{include_summaries}{Logical: should error summaries (obtained via \code{\link[KuskoHarvUtils:get_errors]{KuskoHarvUtils::get_errors()}}) be displayed on the plot?}

\item{period_legend}{Logical: if points are colored by period, should the legend be displayed?
Defaults to \code{TRUE} if \code{period} is not \code{NULL}; ignored if \code{period} is \code{NULL}.}
}
\description{
Plot Predicted vs. Observed Points
}
