% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{predict_model_avg}
\alias{predict_model_avg}
\title{Obtain Model-Averaged Predicted Value}
\usage{
predict_model_avg(fit_list, newdata)
}
\arguments{
\item{fit_list}{\code{\link[base:list]{list}} of fitted model objects of class \code{\link[stats:lm]{lm}}, generally the output of \code{\link[=fit_all_subsets_one]{fit_all_subsets_one()}}.}

\item{newdata}{\code{\link[base:data.frame]{data.frame}} storing the variable values at which to obtain model-averaged predictions.}
}
\value{
Numeric vector of length equal to \code{nrow(newdata)}.
}
\description{
Obtain Model-Averaged Predicted Value
}
