% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{predict_sonar}
\alias{predict_sonar}
\title{Predict Sonar Statistics based on Percentiles}
\usage{
predict_sonar(date, lag)
}
\arguments{
\item{date}{datetime object storing the date of the opener to be predicted.}

\item{lag}{number of days before the proposed opener the prediction is to be made.}

\item{fit_data}{\code{\link[base:data.frame]{data.frame}} storing the variables for regression fitting; generated by \code{\link[KuskoHarvData:prepare_regression_data]{KuskoHarvData::prepare_regression_data()}}}
}
\value{
A \code{\link[base:data.frame]{data.frame}} object with a single row, and columns for
\code{date}, \code{total_sonar_count}, \code{chinook_sonar_comp}, \code{chum_sonar_comp}, and \code{sockeye_sonar_comp}, the latter 4 contain the predicted values.
}
\description{
A method for anticipating sonar total count and species composition
variables to use in harvest prediction.
}
\details{
This prediction method works by assuming the percentile of the sonar statistic on the day of the opener
relative to all historical years (since 2018) will be the same as it is on the day of prediction.
E.g., for a proposed June 12 opener with a 5 day lag: if on June 7 it is determined that 25\% of all historical sonar stats fell below
this year's stat, then the prediction for June 12 is obtained as the value that 25\% of all historical sonar stats
fell below on June 12.

This approach was developed in response to peer review feedback, given a concern that our original LOO analysis under-represented
errors in prediction because we originally evaluated them at the actual BTF statistics (which would be unavailable to users of the model -- manuscript version used BTF values rather than sonar).
By using this approach instead, the intention is to use a more realistic version of what will actually be available to users of the
tool for evaluation of predictive performance.
The notion extends to the use of sonar in current and future versions of the analysis.
}
\note{
This function may only be used if \code{date} is contained within the data set \code{sonar_data_all} in the \code{KuskoHarvData} package.
}
