% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_predictive_tool}
\alias{run_predictive_tool}
\title{Start a Shiny Server for the Prediction Tool and Render It}
\usage{
run_predictive_tool()
}
\description{
Start a Shiny Server for the Prediction Tool and Render It
}
\note{
The tool can be accessed via \href{https://bstaton.shinyapps.io/KuskoHarvPred-tool/}{URL},
however, please use \code{run_predictive_tool()}if you plan to use it often.
}
