% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpercentize}
\alias{unpercentize}
\title{Convert a Character Percent to Numeric Proportion}
\usage{
unpercentize(x)
}
\arguments{
\item{x}{Character, formatted like \code{"XX\%"}}
}
\description{
Convert a Character Percent to Numeric Proportion
}
