% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{whole_loo_analysis}
\alias{whole_loo_analysis}
\title{Conduct the Entire Leave-one-Out Analysis}
\usage{
whole_loo_analysis(
  global_formulae,
  fit_data,
  var_desc = NULL,
  error_types = c("RHO", "ME", "MAE", "MPE", "MAPE"),
  ...
)
}
\arguments{
\item{global_formulae}{A \code{\link[base:list]{list}} object with elements for \code{effort}, \code{totalcpt}, \code{chinook_comp}, \code{chum_comp}, and \code{sockeye_comp} to be passed to the
\code{formula} argument of \code{\link[=fit_global_model_one]{fit_global_model_one()}} separately.}

\item{var_desc}{Optional character string describing the variables used in fitting.
If supplied, will become the first column in the \code{error_summary} element of the output list.}

\item{error_types}{A character vector specifying which types of error summaries (obtained by \code{\link[KuskoHarvUtils:get_errors]{KuskoHarvUtils::get_errors()}}) to return.
Options are any combination of \code{"RHO"}, \code{"RMSE"}, \code{"ME"}, \code{"MAE"}, \code{"MPE"}, or \code{"MAPE"}.}

\item{...}{Optional arguments passed to \code{\link[=fit_all_subsets]{fit_all_subsets()}}}
}
\value{
A \code{\link[base:list]{list}} object with elements:
\itemize{
\item \code{error_summary}: a \code{\link[base:list]{list}} object with elements storing \code{\link[base:data.frame]{data.frame}} objects, where each data frame stores
info about the settings of the run (the values of \code{reduce_colinearity}, \code{cwt_retain}, and \code{var_desc}) and the period- and response variable-specific error summaries.
\item \code{elapsed}: a \code{\link[base:data.frame]{data.frame}} object similar to the \code{error_summary} element, except with a column for minutes elapsed.
\item \code{loo_preds}: a \code{\link[base:data.frame]{data.frame}} object with the model-averaged leave-one-out predictions by record and response variable.
\item \code{models}: a \code{\link[base:list]{list}} object with three elements, one each for \code{effort}, \code{cpt}, and \code{comp}. Each element is the \code{\link[base:list]{list}}
of fitted \code{\link[stats:lm]{lm}} model objects returned by \code{\link[=fit_all_subsets]{fit_all_subsets()}}.
}
}
\description{
A wrapper around model-fitting, model-averaging, leave-one-out calculations, and error summaries
for effort, catch rate, and species composition; also returns error summaries for Chinook and chum+sockeye harvest.
}
\details{
This function conducts several steps in a self-contained wrapper:
\enumerate{
\item Fits the global models to each response variable using \code{\link[=fit_global_model_one]{fit_global_model_one()}}.
\item Fits all subsets of these global models using \code{\link[=fit_all_subsets]{fit_all_subsets()}}.
\item Obtains model-averaged leave-one-out predictions for each record using \code{\link[=loo_pred_model_avg]{loo_pred_model_avg()}}.
\item Summarizes the errors made in step 3 by response variable type and period (see \code{\link[KuskoHarvUtils:get_errors]{KuskoHarvUtils::get_errors()}} and \code{\link[KuskoHarvUtils:get_period]{KuskoHarvUtils::get_period()}}).
}
}
