% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_variables}
\alias{find_variables}
\title{Find Predictor Variable Names in a Set of Fitted Models}
\usage{
find_variables(fit_list)
}
\arguments{
\item{fit_list}{List of fitted model objects of class \code{\link[stats:lm]{lm}}}
}
\description{
Find Predictor Variable Names in a Set of Fitted Models
}
