% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mape}
\alias{get_mape}
\title{Extract a Period- and Variable-specific MAPE value}
\usage{
get_mape(response, period)
}
\arguments{
\item{response}{Character; one of \code{"effort"}, \code{"total_cpt"}, \code{"chinook_comp"}, \code{"chum_comp"}, \code{"sockeye_comp"}, \code{"chinook_harv"}, \code{"chum_harv"}, or \code{"sockeye_harv"}}

\item{period}{Numeric; one of \code{1}, \code{2}, or \code{3}.}
}
\description{
Extract a Period- and Variable-specific MAPE value
}
\note{
Queries the \code{KuskoHarvPred:::loo_output$error_summary} table
for the correct value. The value is on the proportional error scale.
}
