% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{loo_pred}
\alias{loo_pred}
\title{Obtain Leave-one-Out Predictions from a Single Model}
\usage{
loo_pred(fit)
}
\arguments{
\item{fit}{A fitted model object of class \code{\link[stats:lm]{lm}}.}
}
\value{
A \code{\link[base:matrix]{matrix}} object with columns storing values for each record and rows storing the predicted value for
the left out observation and the AICc value of the model fitted after leaving the data point out.
}
\description{
Obtain Leave-one-Out Predictions from a Single Model
}
