% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fitting.R
\name{predict_model_avg}
\alias{predict_model_avg}
\title{Obtain Model-Averaged Predicted Value}
\usage{
predict_model_avg(fit_list, newdata)
}
\arguments{
\item{fit_list}{List of fitted model objects of class \code{\link[stats:lm]{lm}}.}

\item{newdata}{Data frame storing the variable values at which to obtain model-averaged predictions.}
}
\value{
Numeric vector of length equal to \code{nrow(newdata)}.
}
\description{
Obtain Model-Averaged Predicted Value
}
