% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{relationship_plot}
\alias{relationship_plot}
\title{Plot the Relationship of a Variable with Day of the Season}
\usage{
relationship_plot(
  response,
  settings = list(),
  separate_day_types = TRUE,
  pred_day = NULL,
  pred_response = NULL,
  draw_mape_range = FALSE
)
}
\arguments{
\item{response}{Character; one of \code{"effort"}, \code{"total_cpt"}, \code{"chinook_comp"}, \code{"chum_comp"}, or \code{"sockeye_comp"}}

\item{settings}{List specifying which covariate settings to subset the predicted values for. Passed to \code{\link[=subset_pred_data]{subset_pred_data()}}.}

\item{separate_day_types}{Logical; if the variable passed to \code{response} used the predictor variable \code{not_first_day}, should two relationships be drawn?}

\item{pred_day}{Numeric; the day corresponding to a hypothetical prediction (\code{pred_response}). Defaults to \code{NULL} in which case this is not drawn.}

\item{pred_response}{Numeric; the predicted response corresponding to a hypothetical day (\code{pred_day}). Defaults to \code{NULL} in which case this is not drawn.}

\item{draw_make_range}{Logical; should a shaded region around model prediction that shows +/- 1MAPE be shown?}
}
\description{
Plot the Relationship of a Variable with Day of the Season
}
