% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_predictive_tool}
\alias{run_predictive_tool}
\title{Start a Shiny Server for the Prediction Tool and Render It}
\usage{
run_predictive_tool()
}
\description{
Start a Shiny Server for the Prediction Tool and Render It
}
