% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{subset_pred_data}
\alias{subset_pred_data}
\title{Subset Pre-Processed Model Predictions}
\usage{
subset_pred_data(response, settings = list())
}
\arguments{
\item{response}{Character; one of \code{"effort"}, \code{"total_cpt"}, \code{"chinook_comp"}, \code{"chum_comp"}, or \code{"sockeye_comp"}}

\item{settings}{List specifying which covariate settings to subset the predicted values for. See details.}
}
\description{
Subset Pre-Processed Model Predictions
}
\details{
The \code{settings} argument must be a list, and if it is empty (the default), the acceptable elements will be populated with preset values.
The following elements are acceptable:
* \code{settings$day}: numeric; defaults to \code{12:46}
* \code{settings$hours_open}: numeric; must contain any combination of 6, 12, 18, 24, defaults to 12
* \code{settings$p_before_noon}: numeric; must contain any combination of 0.25, 0.5, 0.75, 1; defaults to 0.5
* \code{settings$fished_yesterday}: logical; defaults to \code{FALSE}
* \code{settings$weekend}: logical; defaults to \code{FALSE}
* \code{settings$CAT_total_btf_cpue}: character, must contain any combination of \code{"q10"}, \code{"q25"}, \code{"q50"}, \code{"q75"}, \code{"q90"}; defaults to \code{"q50"}
* \code{settings$CAT_chinook_btf_comp}: same as \code{settings$CAT_total_btf_cpue}
* \code{settings$CAT_chum_btf_comp}: same as \code{settings$CAT_total_btf_cpue}
* \code{settings$CAT_sockeye_btf_comp}: same as \code{settings$CAT_total_btf_cpue}
* \code{settings$CAT_mean_Nwind}: character, must contain any combination of \code{"strong_northerly"}, \code{"none"}, \code{"strong_southerly"}; defaults to \code{"none"}
* \code{settings$CAT_mean_Ewind}: character, must contain any combination of \code{"strong_easterly"}, \code{"none"}, \code{"strong_westerly"}; defaults to \code{"none"}
@note Because not every predictor variable is used for all responses, it is possible to change the \code{settings} argument and receive the same output.
}
