% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vars_biplot}
\alias{vars_biplot}
\title{Create a Scatterplot of Two Variables}
\usage{
vars_biplot(
  xvar = "total_btf_cpue",
  yvar = "total_cpt",
  color_periods = FALSE,
  label_years = FALSE,
  dat = KuskoHarvData::prepare_regression_data()
)
}
\arguments{
\item{xvar}{Character; name of variable to show on the x-axis, defaults to \code{"total_btf_cpue"}}

\item{yvar}{Character; name of variable to show on the y-axis, defaults to \code{"total_cpt"}}

\item{color_periods}{Logical; should points be color-coded to show which time period of the season the data came from?
See \code{\link[KuskoHarvUtils:get_period]{KuskoHarvUtils::get_period()}}. Defaults to \code{FALSE}.}

\item{label_years}{Logical; should points be labeled to show which year the data came from?}

\item{dat}{Data frame; the input regression data set, defaults to \code{\link[KuskoHarvData:prepare_regression_data]{KuskoHarvData::prepare_regression_data()}}.}
}
\description{
For investigating basic co-linearity in predictors,
relationships among variables not available
from \code{\link[=relationship_plot]{relationship_plot()}}.
}
